+++ src/Models/WorkflowInterrupt.php
@@ -19,6 +19,7 @@ use TimMcLeod\AgentWorkflows\Enums\InterruptType;
  * @property string|null $resolved_by_type
  * @property int|string|null $resolved_by_id
  * @property array<string, mixed>|null $resolution
+ * @property Carbon|null $timeout_at
  * @property Carbon|null $resolved_at
  * @property Carbon $created_at
  * @property Carbon $updated_at
@@ -39,6 +40,7 @@ class WorkflowInterrupt extends Model
             'response_schema' => 'array',
             'context' => 'array',
             'resolution' => 'array',
+            'timeout_at' => 'datetime',
             'resolved_at' => 'datetime',
         ];
     }